+++ app/src/main/java/com/example/daytask/ui/screens/newtask/ClearInputField.kt
@@ -7,11 +7,9 @@ import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusDirection
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.dimensionResource
@@ -22,8 +20,9 @@ import com.example.daytask.R
 import com.example.daytask.ui.theme.HelpText
 import com.example.daytask.ui.theme.InputText
 import com.example.daytask.ui.theme.PlaceholderColor
-import com.example.daytask.ui.theme.Secondary
 import com.example.daytask.ui.theme.White
+import com.example.daytask.util.TextFieldManager
+import com.example.daytask.util.TextFieldManager.clearFocusOnKeyboardDismiss
 
 @Composable
 fun ClearInputField(
@@ -50,16 +49,7 @@ fun ClearInputField(
             value = inputText,
             onValueChange = onValueChange,
             textStyle = textStyle,
-            colors = TextFieldDefaults.colors(
-                focusedContainerColor = Secondary,
-                unfocusedContainerColor = Secondary,
-                focusedLeadingIconColor = Color.Unspecified,
-                unfocusedLeadingIconColor = Color.Unspecified,
-                focusedTrailingIconColor = Color.Unspecified,
-                unfocusedTrailingIconColor = Color.Unspecified,
-                focusedIndicatorColor = Color.Transparent,
-                unfocusedIndicatorColor = Color.Transparent
-            ),
+            colors = TextFieldManager.colors(),
             shape = RectangleShape,
             placeholder = {
                 Text(
@@ -69,7 +59,9 @@ fun ClearInputField(
             },
             keyboardOptions = keyboardOptions,
             keyboardActions = keyboardActions,
-            modifier = Modifier.fillMaxWidth()
+            modifier = Modifier
+                .fillMaxWidth()
+                .clearFocusOnKeyboardDismiss()
         )
     }
 }
