+++ app/src/main/java/com/example/daytask/ui/screens/newtask/NewTaskViewModel.kt
@@ -5,12 +5,12 @@ import androidx.lifecycle.ViewModel
 import com.example.daytask.R
 import com.example.daytask.data.Task
 import com.example.daytask.data.User
-import com.example.daytask.util.Constants.tenMinutesInMillis
+import com.example.daytask.ui.screens.tools.FirebaseManager
+import com.example.daytask.util.Constants.timeLimitMillis
 import com.example.daytask.util.NetworkManager.isNetworkAvailable
 import com.example.daytask.util.NotifyManager.notifyUser
 import com.example.daytask.util.Status
 import com.google.firebase.auth.ktx.auth
-import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -20,10 +20,6 @@ import java.util.Calendar
 class NewTaskViewModel : ViewModel() {
     private val _uiState = MutableStateFlow(NewTaskUiState())
     val uiState = _uiState.asStateFlow()
-
-    private val database = Firebase.database.reference
-    private val userId = Firebase.auth.currentUser!!.uid
-    private val ref = database.child("users/$userId/tasks").push()
 
     fun updateUiState(uiState: NewTaskUiState) = _uiState.update { uiState }
 
@@ -58,7 +54,7 @@ class NewTaskViewModel : ViewModel() {
             subTasksList = listOf(),
             taskComplete = false
         )
-        ref.setValue(newTask)
+        FirebaseManager.uploadTask(Firebase.auth.currentUser!!.uid, newTask)
             .addOnCompleteListener { task ->
                 if (task.isSuccessful) {
                     updateUiState(_uiState.value.copy(updateResult = true))
@@ -71,7 +67,7 @@ class NewTaskViewModel : ViewModel() {
 data class NewTaskUiState(
     val title: String = "",
     val details: String = "",
-    val date: Long = Calendar.getInstance().timeInMillis + tenMinutesInMillis,
+    val date: Long = Calendar.getInstance().timeInMillis + timeLimitMillis,
     val memberList: List<User> = listOf(),
     val status: Status = Status.Done,
     val updateResult: Boolean = false
