+++ app/src/main/java/com/example/daytask/ui/screens/tools/InputField.kt
@@ -10,14 +10,12 @@ import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.painterResource
@@ -30,8 +28,9 @@ import com.example.daytask.R
 import com.example.daytask.ui.theme.HelpText
 import com.example.daytask.ui.theme.InputText
 import com.example.daytask.ui.theme.PlaceholderColor
-import com.example.daytask.ui.theme.Secondary
 import com.example.daytask.ui.theme.White
+import com.example.daytask.util.TextFieldManager
+import com.example.daytask.util.TextFieldManager.clearFocusOnKeyboardDismiss
 
 @Composable
 fun InputField(
@@ -99,16 +98,7 @@ fun InputField(
         textStyle = InputText.copy(White),
         keyboardOptions = keyboardOptions,
         keyboardActions = keyboardActions,
-        colors = TextFieldDefaults.colors(
-            focusedContainerColor = Secondary,
-            unfocusedContainerColor = Secondary,
-            focusedLeadingIconColor = Color.Unspecified,
-            unfocusedLeadingIconColor = Color.Unspecified,
-            focusedTrailingIconColor = Color.Unspecified,
-            unfocusedTrailingIconColor = Color.Unspecified,
-            focusedIndicatorColor = Color.Transparent,
-            unfocusedIndicatorColor = Color.Transparent
-        ),
+        colors = TextFieldManager.colors(),
         shape = RectangleShape,
         placeholder = {
             Text(
@@ -129,6 +119,8 @@ fun InputField(
         } else VisualTransformation.None,
         isError = isError,
         supportingText = if (isError) supportingTextView else null,
-        modifier = modifier.fillMaxWidth()
+        modifier = modifier
+            .fillMaxWidth()
+            .clearFocusOnKeyboardDismiss()
     )
 }
